scripts/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
audiomentations==0.43.1
audioread==3.0.1
bitsandbytes==0.44.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
datasets==3.0.2
decorator==5.2.1
dill==0.3.8
evaluate==0.4.3
filelock==3.13.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.6.1
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.11
Jinja2==3.1.4
jiwer==4.0.0
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.3
numba==0.62.1
numpy==2.1.2
numpy-minmax==0.5.0
numpy-rms==0.6.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==25.0
pandas==2.2.3
peft==0.13.2
pillow==11.0.0
platformdirs==4.5.0
pooch==1.8.2
propcache==0.4.1
psutil==7.1.0
pyarrow==21.0.0
pycparser==2.23
pyparsing==3.2.5
pythainlp==5.1.2
python-dateutil==2.9.0.post0
python-stretch==0.3.1
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.44.2
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
